# docs/openapi_spec.yaml
openapi: 3.1.0
info:
  title: Sistema RAG — Catálogo de Flores y Listas de Precios
  version: 1.0.0
  summary: Especificación formal del backend FastAPI del sistema RAG del proyecto final.
  description: |
    Contrato HTTP del sistema RAG. **Esta especificación es la fuente de verdad**
    y precede a cualquier código de implementación (`app/`) y a las pruebas (`tests/`).

    ## Capas y responsabilidades (arquitectura obligatoria de la rúbrica)

    | Capa | Tecnología | Responsabilidad | Puede tocar |
    |---|---|---|---|
    | UI | Streamlit (`:8501`) | Cargar documentos, preguntar, mostrar respuesta + citas `[n]` + scores | **solo** HTTP a FastAPI |
    | API | FastAPI (`:8000`) | Orquestar chunking, embeddings, k-NN y generación | Google AI, ChromaDB |
    | Índice | ChromaDB (`./chroma`) | Persistencia de chunks + vectores, búsqueda k-NN | solo la API |
    | Embeddings | Google AI (`google-genai`) | Vectorizar chunks **y** preguntas con el **mismo** modelo |_VECTORIZA. No genera texto_ |
    | Generación | Gemini (`google-genai`) | Redactar respuesta anclada en la evidencia o abstenerse | — |

    ## Reglas de diseño que esta especificación deja fijadas

    1. **Un único proveedor de embeddings.** `EMBEDDING_MODEL` incrusta documentos y
       preguntas. Está prohibido el embedder por defecto de ChromaDB: la API siempre
       envía `embeddings=` calculados por Google AI.
    2. **Recuperar antes de generar.** `/query` ejecuta siempre k-NN sobre ChromaDB y
...
servers:
...
security: [] # Sin autenticación: la clave de Google AI vive solo en el entorno del servidor.
paths:
...
  /ingest:
    post:
      tags: [ingesta]
      operationId: ingestDocuments
      summary: Chunkifica, incrusta con Google AI y persiste en ChromaDB.
      description: |
        **Todos los vectores los produce Google AI.** El flujo es:
        extracción de texto → `chunk.py` (trozos con solape) → `embed.py`
        (lotes pequeños) → `store.py` (`add` con `embeddings=` propios).

        Acepta **cualquiera** de los dos modos declarados en `content`, y ambos
        pueden combinarse en la misma petición:

        * `multipart/form-data` — archivos subidos desde Streamlit (`files`).
        * `application/json` — rutas dentro del servidor (`paths`), relativas a
          `data/` por defecto.

        Extensiones admitidas: `.md`, `.txt`, `.pdf` (con `pypdf`), `.csv`.
        PDF escaneado sin capa de texto no se indexa: se cuenta en
        `skipped` con motivo `sin_capa_de_texto` para no falsear el corpus.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngestRequest' }
            examples:
              rutas_del_servidor:
                summary: Indexar la carpeta data/ del proyecto
                value:
                  paths: ['data/']
                  reset: false
              parametros_de_chunking:
                summary: Forzar tamaño y solape distintos a los del entorno
                value:
                  paths: ['data/catalogo_rosas.md', 'data/precios_2026.csv']
                  chunk_size: 320
                  chunk_overlap: 64
                  reset: true
          multipart/form-data:
            schema: { $ref: '#/components/schemas/IngestUploadRequest' }
            encoding:
              files:
                contentType: text/markdown, text/plain, application/pdf, text/csv
            examples:
              carga_desde_streamlit:
                summary: Formulario de subida de la UI
                value:
                  files: ['catalogo_rosas.md', 'lista_precios.csv']
                  reset: false
      responses:
        '200':
          description: Ingesta completada. Siempre `200`, aunque algunos archivos se omitan (mira `skipped`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResponse' }
              examples:
                exito:
                  summary: 5 documentos, 412 chunks
                  value:
                    documents_processed: 5
                    chunks_indexed: 412
                    chunks_skipped: 0
                    collection: flowers_catalog
                    embedding_model: gemini-embedding-001
                    sources:
                      - data/catalogo_rosas.md
                      - data/catalogo_tulipanes.md
                      - data/precios_2026.csv
                      - data/cuidado_iris.md
                      - data/envio_y_garantias.md
                    skipped: []
                    duration_ms: 21430
        '400':
          description: No se recibió ningún archivo ni ruta legible.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                sin_entrada:
                  value:
                    detail: 'No se recibió ningún archivo ni ruta. Usa `files` (subida) o `paths` (rutas del servidor).'
                    code: ingest_no_input
                    hint: null
        '413':
          description: Algún archivo supera `max_file_bytes` (25 MB por defecto).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                archivo_grande:
                  value:
                    detail: 'El archivo `catalogo.pdf` supera el límite de 26214400 bytes.'
                    code: file_too_large
                    hint: 'Divide el PDF o eleva MAX_FILE_BYTES en el entorno.'
        '503':
          description: Falta `GOOGLE_API_KEY` o ChromaDB no está accesible; no se puede indexar.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                sin_clave:
                  value:
                    detail: 'GOOGLE_API_KEY no está configurada en el servidor.'
                    code: missing_google_api_key
                    hint: 'Copia .env.example a .env y define GOOGLE_API_KEY (https://aistudio.google.com/apikey).'
                chroma_caido:
                  value:
                    detail: 'ChromaDB no está accesible en ./chroma.'
                    code: chroma_unavailable
                    hint: null
  /query:
    post:
      tags: [consulta]
      operationId: queryRag
      summary: Recupera top-k chunks en ChromaDB y genera una respuesta anclada con Gemini.
      description: |
        Orden obligatorio: (1) incrustar la pregunta con `EMBEDDING_MODEL`,
        (2) k-NN por coseno en ChromaDB, (3) filtro `MIN_SCORE`, (4) **solo
        entonces** Gemini responde con los chunks numerados `[1]`, `[2]`, …
        En español, citando `[n]`, y absteniéndose si la evidencia no cubre.

        **Regla de abstención (dos capas, ambas visibles en la respuesta):**
        * Capa determinista — si ningún vecino alcanza `min_score`, o `top_k` se resuelve a 0:
          `abstained: true`, `abstain_reason: "sin_evidencia"`, sin llamada a Gemini y
          `citations: []`. El índice vacío no es una abstención sino el `503 empty_index`
          declarado abajo: la UI necesita distinguir "no hay nada indexado" de
          "no hay evidencia para esta pregunta".
        * Capa del modelo — si la evidencia no cubre la pregunta, Gemini devuelve
          la frase de abstención acordada y la API marca `reason: "modelo_abstuvo"`
          con las citas que sí se recuperaron.

        El cliente (Streamlit) **nunca** recibe `500` por una pregunta fuera de
        dominio: el resultado de abstención es un `200` de negocio.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryRequest' }
            examples:
              pregunta_del_dominio:
                summary: Consulta que sí tiene evidencia
                value:
                  question: '¿Cuál es el precio de un ramo de 20 rosas rojas y si incluye IVA?'
                  top_k: 3
              pregunta_fuera_de_dominio:
                summary: Impossible de responder con el corpus → debe abstenerse
                value:
                  question: '¿Quién ganó el Mundial de fútbol de 1998 y con qué margen?'
                  top_k: 3
              con_filtro_de_fuente:
                summary: Restringir la búsqueda a un documento (reto opcional)
                value:
                  question: '¿Cada cuánto se riegan las orquídeas?'
                  top_k: 4
                  source: data/cuidado_iris.md
                  min_score: 0.45
      responses:
        '200':
          description: Respuesta RAG o abstención (ambos son éxito de negocio).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryResponse' }
              examples:
                con_evidencia:
                  summary: Respuesta anclada con citas
                  value:
                    answer: 'El ramo de 20 rosas rojas cuesta 89,90 € con IVA incluido [1]. El envío es gratuito a partir de 60 € [2].'
                    abstained: false
                    abstain_reason: null
                    citations:
                      - index: 1
                        id: 'data/catalogo_rosas.md#chunk-004'
                        source: data/catalogo_rosas.md
                        title: Catálogo de rosas
                        chunk_index: 4
                        page: null
                        text: 'Ramo de 20 rosas rojas — 89,90 € (IVA incluido).'
                        score: 0.78
                      - index: 2
                        id: 'data/envio_y_garantias.md#chunk-001'
                        source: data/envio_y_garantias.md
                        title: Envío y garantías
                        chunk_index: 1
                        page: null
                        text: 'Envío gratuito para pedidos superiores a 60 € en península.'
                        score: 0.61
                    question_embedding_model: gemini-embedding-001
                    generation_model: gemini-3.5-flash
                    top_k: 3
                    retrieval_ms: 412
                    generation_ms: 1180
                abstain_por_score:
                  summary: Ningún chunk alcanza min_score → abstención sin llamar a Gemini
                  value:
                    answer: 'No tengo evidencia suficiente en el corpus indexado para responder a esa pregunta.'
                    abstained: true
                    abstain_reason: sin_evidencia
                    citations: []
                    question_embedding_model: gemini-embedding-001
                    generation_model: null
                    top_k: 3
                    retrieval_ms: 380
                    generation_ms: 0
                abstain_por_modelo:
                  summary: Hay chunks parecidos pero no responden la pregunta
                  value:
                    answer: 'No encuentro en los documentos cargados información que responda a esa pregunta.'
                    abstained: true
                    abstain_reason: modelo_abstuvo
                    citations:
                      - index: 1
                        id: 'data/catalogo_rosas.md#chunk-002'
                        source: data/catalogo_rosas.md
                        title: Catálogo de rosas
                        chunk_index: 2
                        page: null
                        text: 'Las rosas se conservan entre 3 y 5 días en agua fresca.'
                        score: 0.52
                    question_embedding_model: gemini-embedding-001
                    generation_model: gemini-3.5-flash
                    top_k: 3
                    retrieval_ms: 401
                    generation_ms: 940
        '400':
          description: '`question` vacía o solo con espacios.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                vacia:
                  value:
                    detail: 'La pregunta no puede estar vacía.'
                    code: empty_question
                    hint: null
        '503':
          description: Índice vacío, ChromaDB inaccesible, falta `GOOGLE_API_KEY` o fallo de Gemini.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                gemini_fallo:
                  value:
                    detail: 'Google AI falló al generar la respuesta con gemini-3.5-flash.'
                    code: google_ai_error
                    hint: 'Reintenta en unos segundos; si persiste, revisa la cuota en https://ai.dev/rate-limit.'
                indice_vacio:
                  value:
                    detail: 'El índice está vacío. Ingesta documentos primero con POST /ingest.'
                    code: empty_index
                    hint: null
                sin_clave:
                  value:
                    detail: 'GOOGLE_API_KEY no está configurada en el servidor.'
                    code: missing_google_api_key
                    hint: 'Obtén la clave en https://aistudio.google.com/apikey'
        '504':
          description: Google AI no respondió a tiempo (embeddings o generación).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                timeout:
                  value:
                    detail: 'Google AI tardó más de 30 s en responder.'
                    code: google_ai_timeout
                    hint: null
...
components:
  schemas:
...
    IngestRequest:
      type: object
      description: |
        Modo JSON — indexar documentos que ya están en el disco del servidor.
        `paths` no es obligatorio en el esquema para que una petición sin rutas llegue al
        endpoint y devuelva el `400 ingest_no_input` declarado, no un `422` genérico.
      properties:
        paths:
          type: array
          description: Rutas de archivo o carpeta, relativas a la raíz del proyecto (o absolutas).
          items: { type: string }
          minItems: 1
          examples:
            - ['data/']
            - ['data/catalogo_rosas.md', 'data/precios_2026.csv']
        chunk_size:
          type: integer
          minimum: 100
          maximum: 1000
          default: 320
          description: 'Palabras por chunk (rango de la rúbrica: 200–400).'
          example: 320
        chunk_overlap:
          type: integer
          minimum: 0
          description: 'Palabras de solape entre chunks consecutivos (rango de la rúbrica: 40–80). `0` desactiva el solape.'
          default: 64
          example: 64
        reset:
          type: boolean
          default: false
          description: Si `true`, vacía la colección antes de indexar (reindexado limpio).
    IngestUploadRequest:
      type: object
      description: Modo multipart — archivos subidos desde Streamlit.
      required: [files]
      properties:
        files:
          type: array
          minItems: 1
          description: Binarios de los documentos (.md, .txt, .pdf, .csv). Se guardan en `data/uploads/` y se indexan con `source` = esa ruta.
          items:
            type: string
            format: binary
        paths:
          type: array
          description: Rutas adicionales del servidor a incluir en la misma ingesta.
          items: { type: string }
        chunk_size:
          type: integer
          minimum: 100
          maximum: 1000
          default: 320
        chunk_overlap:
          type: integer
          minimum: 0
          default: 64
        reset:
          type: boolean
          default: false
    IngestResponse:
      type: object
      description: Resultado de la ingesta. `chunks_indexed` es el número de vectores de Google AI guardados en ChromaDB.
      required: [documents_processed, chunks_indexed, collection, embedding_model, sources]
      properties:
        documents_processed:
          type: integer
          minimum: 0
          description: Documentos leídos y troceados con éxito.
          example: 5
        chunks_indexed:
          type: integer
          minimum: 0
          description: Chunks (y por tanto embeddings) escritos en la colección.
          example: 412
        chunks_skipped:
          type: integer
          minimum: 0
          default: 0
          description: Chunks no indexados por estar vacíos o superar el límite de tokens del modelo.
          example: 0
        collection:
          type: string
          example: flowers_catalog
        embedding_model:
          type: string
          description: Confirma que los vectores salieron de Google AI, no del embedder por defecto de ChromaDB.
          example: gemini-embedding-001
        sources:
          type: array
          description: Rutas efectivamente indexadas; es la lista válida para el parámetro `source` de `/query`.
          items: { type: string }
          example: ['data/catalogo_rosas.md', 'data/precios_2026.csv']
        skipped:
          type: array
          description: Archivos omitidos con el motivo; la ingesta no es atómica, así que se reportan en lugar de fallar.
          items: { $ref: '#/components/schemas/SkippedFile' }
        duration_ms:
          type: integer
          minimum: 0
          description: Duración total; útil para ver si el cuello es la cuota de la API de Google AI.
          example: 21430
    SkippedFile:
      type: object
      required: [source, reason]
      properties:
        source: { type: string, example: 'data/escaneado.pdf' }
        reason:
          type: string
          enum: [sin_capa_de_texto, formato_no_soportado, vacio, ilegible, demasiado_grande]
          example: sin_capa_de_texto
    QueryRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
          description: Pregunta en lenguaje natural. Se recorta; si queda vacía, `400 empty_question`.
          examples:
            - '¿Cuál es el precio de un ramo de 20 rosas rojas y si incluye IVA?'
            - '¿Quién ganó el Mundial de fútbol de 1998?'
        top_k:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
          description: 'Vecinos a recuperar de ChromaDB. Rango razonable según rúbrica: 2–5.'
          example: 3
        source:
          type: string
          description: 'Filtro opcional por documento (reto opcional): restringe el k-NN a los chunks de ese `source`.'
          example: data/cuidado_iris.md
        min_score:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: |
            Umbral de similitud coseno (score = 1 − distancia coseno) por debajo del cual
            **no hay evidencia**. Default `MIN_SCORE` del entorno. Si ningún vecino lo
            alcanza, la API se abstiene sin llamar a Gemini.
          default: 0.35
          example: 0.35
    QueryResponse:
      type: object
      required: [answer, abstained, abstain_reason, citations]
      properties:
        answer:
          type: string
          description: |
            Texto en español generado por Gemini **solo** con los chunks dados. Debe
            incluir citas `[n]` que correspondan a `citations[index - 1]`. Si
            `abstained` es `true`, dice explícitamente que no hay evidencia suficiente
            y no rellena con conocimiento paramétrico.
          examples:
            - 'El ramo de 20 rosas cuesta 89,90 € con IVA incluido [1].'
            - 'No tengo evidencia suficiente en el corpus indexado para responder a esa pregunta.'
        abstained:
          type: boolean
          description: '`true` = el sistema se abstuvo. Es una respuesta válida, no un error.'
          example: false
        abstain_reason:
          type: [string, 'null']
          description: Motivo de la abstención; `null` cuando `abstained` es `false`.
          enum: [sin_evidencia, modelo_abstuvo, indice_vacio, modelo_sin_clave, error_generacion, 'null']
          examples: [sin_evidencia, modelo_abstuvo, null]
        citations:
          type: array
          description: |
            Chunks recuperados, ordenados por score descendente y **numerados desde 1**
            con la misma `index` que aparece como `[n]` en `answer`. En la abstención por
            score van vacíos; en la abstención por modelo conservan lo recuperado para que
            la UI muestre por qué se abstuvo.
          items: { $ref: '#/components/schemas/Citation' }
        question_embedding_model:
          type: string
          description: Modelo usado para vectorizar la pregunta; debe coincidir con el de los chunks.
          example: gemini-embedding-001
        generation_model:
          type: [string, 'null']
          description: Modelo Gemini usado; `null` si no se llegó a generar (abstención determinista).
          example: gemini-3.5-flash
        top_k:
          type: integer
          description: '`top_k` efectivo tras aplicar el máximo permitido.'
          example: 3
        retrieval_ms:
          type: integer
          minimum: 0
          example: 412
        generation_ms:
          type: integer
          minimum: 0
          example: 1180
...
    Error:
      type: object
      description: Cuerpo de error uniforme de la API. `code` es estable y sirve para que la UI elija el mensaje.
      required: [detail, code]
      properties:
        detail:
          type: string
          description: Mensaje legible en español.
          example: 'GOOGLE_API_KEY no está configurada en el servidor.'
        code:
          type: string
          description: Identificador estable del error.
          enum:
            - empty_question
            - ingest_no_input
            - file_too_large
            - missing_google_api_key
            - chroma_unavailable
            - empty_index
            - google_ai_error
            - google_ai_timeout
            - source_not_found
            - internal_error
          example: missing_google_api_key
        hint:
          type: [string, 'null']
          description: Acción sugerida para el usuario (p. ej. cómo obtener la clave).
          example: 'Obtén la clave en https://aistudio.google.com/apikey'
...